Decompose CharacterCreator.tsx (c2): remaining step components
Second slice of the C-c CharacterCreator decomposition. Strict
mechanical extraction — zero behaviour change, every component /
helper / type moves verbatim with imports adjusted.

Three more step components peel out:

- `steps/AttributesStep.tsx` — AttributesStep + its 9 helpers
  (BoostSourceBlock, FlawChips, BoostSlotPicker, FreeBoostBlock,
  useSourceSlots, parseAncestryFlaws, parseAncestryOrBackgroundSlots,
  parseClassKeyAbility, seedPicks) plus the SlotKind / ParsedSlot /
  SourceDocState types they share.
- `steps/SkillsStep.tsx` — SkillsStep + CORE_SKILLS, TrainedSkillsDoc,
  SkillsDocState, normaliseTrainedSkills.
- `steps/LanguagesStep.tsx` — LanguagesStep + COMMON_LANGUAGES,
  AncestryLanguagesDoc, LanguagesDocState, normaliseAncestryLanguages.

`CharacterCreator.tsx` now contains only the route entry: the root
component, StepNav, and computeIntMod. Down from 2204 LOC pre-c1 →
~1100 LOC after c1 → ~370 LOC after c2 (~83% reduction). The audit's
"cuts cognitive load by 60%" target is comfortably exceeded.

`useCharacterDraft` (the audit's other recommendation) was not
included — see PR description.

Bucket C-c2 of the cleanup audit. C-e (BookBrowser+BookReader
decomposition in dm-tool) is next per the agreed order.
